refactor: validator 도메인 안에서 처리
