--- src/main/java/domain/validator/HeightValidator.java
@@ -1,13 +1,0 @@
-package domain.validator;
-
-
-public class HeightValidator {
-    private static final String NOT_POSITIVE_ERROR_MESSAGE = "[ERROR] 양의 정수만 입력해주세요.";
-    private static final int  HEIGHT_MIN_SIZE = 0;
-
-    public void checkNegativeNumber(int height) {
-        if (height <= HEIGHT_MIN_SIZE) {
-            throw new IllegalArgumentException(NOT_POSITIVE_ERROR_MESSAGE);
-        }
-    }
-}
